+++ ql/FiniteDifferences/tridiagonaloperator.cpp
@@ -39,7 +39,7 @@ namespace QuantLib {
     namespace FiniteDifferences {
 
         TridiagonalOperator::TridiagonalOperator(
-            unsigned int size) {
+            size_t size) {
 
             if (size>=3) {
                 diagonal_      = Array(size);
@@ -106,7 +106,7 @@ namespace QuantLib {
 
             // matricial product
             result[0] = diagonal_[0]*v[0] + aboveDiagonal_[0]*v[1];
-            for (unsigned int j=1;j<=size()-2;j++)
+            for (size_t j=1;j<=size()-2;j++)
                 result[j] = belowDiagonal_[j-1]*v[j-1]+ diagonal_[j]*v[j] +
                     aboveDiagonal_[j]*v[j+1];
             result[size()-1] = belowDiagonal_[size()-2]*v[size()-2] +
@@ -175,7 +175,7 @@ namespace QuantLib {
             QL_REQUIRE(bet != 0.0,
                 "TridiagonalOperator::solveFor: division by zero");
             result[0] = bcRhs[0]/bet;
-            unsigned int j;
+            size_t j;
             for (j=1;j<=size()-1;j++){
                 tmp[j]=aboveDiagonal_[j-1]/bet;
                 bet=diagonal_[j]-belowDiagonal_[j-1]*tmp[j];
@@ -183,7 +183,7 @@ namespace QuantLib {
                     "TridiagonalOperator::solveFor: division by zero");
                 result[j] = (bcRhs[j]-belowDiagonal_[j-1]*result[j-1])/bet;
             }
-            // cannot be j>=0 with unsigned int j
+            // cannot be j>=0 with size_t j
             for (j=size()-2;j>0;j--)
                 result[j] -= tmp[j+1]*result[j+1];
             result[0] -= tmp[1]*result[1];
@@ -191,7 +191,7 @@ namespace QuantLib {
             return result;
         }
 
-        TridiagonalOperator TridiagonalOperator::identity(unsigned int size){
+        TridiagonalOperator TridiagonalOperator::identity(size_t size){
             return TridiagonalOperator(
                 Array(size-1, 0.0),     // lower diagonal
                 Array(size,   1.0),     // diagonal
